Fix test.
