+++ cmd/algorand-indexer/daemon_test.go
@@ -3,8 +3,6 @@ package main
 import (
 	"context"
 	"errors"
-	"io"
-	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -12,13 +10,11 @@ import (
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/algorand/go-algorand-sdk/encoding/json"
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/ledger/ledgercore"
 	"github.com/algorand/go-algorand/rpcs"
 
 	"github.com/algorand/indexer/processor/blockprocessor"
-	"github.com/algorand/indexer/util"
 	itest "github.com/algorand/indexer/util/test"
 )
 
@@ -75,33 +71,4 @@ func TestImportRetryAndCancel(t *testing.T) {
 	// Wait for handler to exit.
 	cancel()
 	wg.Wait()
-}
-
-func TestReadGenesis(t *testing.T) {
-	var reader io.Reader
-	// nil reader
-	_, err := util.ReadGenesis(reader)
-	assert.Contains(t, err.Error(), "readGenesis() err: reader is nil")
-	// no match struct field
-	genesisStr := "{\"version\": 2}"
-	reader = strings.NewReader(genesisStr)
-	_, err = util.ReadGenesis(reader)
-	assert.Contains(t, err.Error(), "json decode error")
-
-	genesis := bookkeeping.Genesis{
-		SchemaID:    "1",
-		Network:     "test",
-		Proto:       "test",
-		RewardsPool: "AAAA",
-		FeeSink:     "AAAA",
-	}
-
-	// read and decode genesis
-	reader = strings.NewReader(string(json.Encode(genesis)))
-	_, err = util.ReadGenesis(reader)
-	assert.Nil(t, err)
-	// read from empty reader
-	_, err = util.ReadGenesis(reader)
-	assert.Contains(t, err.Error(), "readGenesis() err: EOF")
-
 }
